use decimal

+++ src/api/account.rs
@@ -1,8 +1,7 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
-use crate::string_as_f64;
-
-use super::{trade::PositionSide, ApiQuery, QueryType};
+use super::{ApiQuery, QueryType, trade::PositionSide};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 #[serde(rename_all = "camelCase")]
@@ -18,28 +17,28 @@ impl ApiQuery for QV2FutureAccountStatus {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct FutureAccountStatus {
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_initial_margin: f64,
-    #[serde(rename = "totalMaintMargin", deserialize_with = "string_as_f64")]
-    pub total_main_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_wallet_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_unrealized_profit: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_margin_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_position_initial_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_open_order_initial_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_cross_wallet_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub total_cross_un_pnl: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub available_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub max_withdraw_amount: f64,
+    #[serde()]
+    pub total_initial_margin: Decimal,
+    #[serde(rename = "totalMaintMargin")]
+    pub total_main_margin: Decimal,
+    #[serde()]
+    pub total_wallet_balance: Decimal,
+    #[serde()]
+    pub total_unrealized_profit: Decimal,
+    #[serde()]
+    pub total_margin_balance: Decimal,
+    #[serde()]
+    pub total_position_initial_margin: Decimal,
+    #[serde()]
+    pub total_open_order_initial_margin: Decimal,
+    #[serde()]
+    pub total_cross_wallet_balance: Decimal,
+    #[serde()]
+    pub total_cross_un_pnl: Decimal,
+    #[serde()]
+    pub available_balance: Decimal,
+    #[serde()]
+    pub max_withdraw_amount: Decimal,
     pub assets: Vec<Asset>,
     pub positions: Vec<Position>,
 }
@@ -48,28 +47,28 @@ pub struct FutureAccountStatus {
 #[serde(rename_all = "camelCase")]
 pub struct Asset {
     pub asset: String,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub wallet_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub unrealized_profit: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub margin_balance: f64,
-    #[serde(rename = "maintMargin", deserialize_with = "string_as_f64")]
-    pub main_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub initial_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub position_initial_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub open_order_initial_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub cross_wallet_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub cross_un_pnl: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub available_balance: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub max_withdraw_amount: f64,
+    #[serde()]
+    pub wallet_balance: Decimal,
+    #[serde()]
+    pub unrealized_profit: Decimal,
+    #[serde()]
+    pub margin_balance: Decimal,
+    #[serde(rename = "maintMargin")]
+    pub main_margin: Decimal,
+    #[serde()]
+    pub initial_margin: Decimal,
+    #[serde()]
+    pub position_initial_margin: Decimal,
+    #[serde()]
+    pub open_order_initial_margin: Decimal,
+    #[serde()]
+    pub cross_wallet_balance: Decimal,
+    #[serde()]
+    pub cross_un_pnl: Decimal,
+    #[serde()]
+    pub available_balance: Decimal,
+    #[serde()]
+    pub max_withdraw_amount: Decimal,
     pub update_time: i64,
 }
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -77,20 +76,20 @@ pub struct Asset {
 pub struct Position {
     pub symbol: String,
     pub position_side: PositionSide,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub position_amt: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub unrealized_profit: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub isolated_margin: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub notional: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub isolated_wallet: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub initial_margin: f64,
-    #[serde(rename = "maintMargin", deserialize_with = "string_as_f64")]
-    pub main_margin: f64,
+    #[serde()]
+    pub position_amt: Decimal,
+    #[serde()]
+    pub unrealized_profit: Decimal,
+    #[serde()]
+    pub isolated_margin: Decimal,
+    #[serde()]
+    pub notional: Decimal,
+    #[serde()]
+    pub isolated_wallet: Decimal,
+    #[serde()]
+    pub initial_margin: Decimal,
+    #[serde(rename = "maintMargin")]
+    pub main_margin: Decimal,
     pub update_time: i64,
 }
 
@@ -137,22 +136,22 @@ pub struct AccountStatus {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct CommissionRate {
-    #[serde(deserialize_with = "string_as_f64")]
-    pub maker: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub taker: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub buyer: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub seller: f64,
+    #[serde()]
+    pub maker: Decimal,
+    #[serde()]
+    pub taker: Decimal,
+    #[serde()]
+    pub buyer: Decimal,
+    #[serde()]
+    pub seller: Decimal,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Balance {
     pub asset: String,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub free: f64,
-    #[serde(deserialize_with = "string_as_f64")]
-    pub locked: f64,
+    #[serde()]
+    pub free: Decimal,
+    #[serde()]
+    pub locked: Decimal,
 }
